+++ test/dummy/config/application.rb
@@ -1,8 +1,9 @@
-require File.expand_path('../boot', __FILE__)
+# frozen_string_literal: true
+
+require File.expand_path("boot", __dir__)
 
 require "action_controller/railtie"
 require "action_mailer/railtie"
-require "sprockets/railtie"
 
 Bundler.require(*Rails.groups)
 require "burgundy"
@@ -11,4 +12,3 @@ module Dummy
   class Application < Rails::Application
   end
 end
-